// app/pro-plus/model/item-list-response.ts
/**
 * Beacon New Rest Service Ver 1.0
 * Release notes:  `4.8`  * Add `onHold` to request body of endpoint `/submitOrder`
 *
 * OpenAPI spec version: release/4.8
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */ import { Category } from './category';
import { ProductListItem } from './product-item-list';

export interface ItemListResponseV2 {
    message?: string;
    totalNumRecs?: number;
    pageNumRecs?: number;
    items?: Array<ProductListItem>;
    facets?: {
        [index: string]: {
            facetName: string;
            facetId: string;
            recordCount: number;
            selected: boolean;
        }[];
    };
    categories?: Array<Category>;
}